Migrated rigid body (kinamatic).

+++ windows_base_plugin_game/wbp_d3d12/include/window_d3d12_facade.h
@@ -59,7 +59,7 @@ namespace wbp_d3d12
 
         virtual ID3D12CommandAllocator *GetCommandAllocator() override;
         virtual ID3D12GraphicsCommandList *GetCommandList() override;
-        virtual const size_t &GetCurrentFrameIndex() const override;
+        virtual const UINT &GetCurrentFrameIndex() const override;
 
         virtual void ClearViews(const float (&clearColor)[4], UINT depthStencilIndex) override;
         virtual void Present() override;
